requirements.txt
fastapi
uvicorn
python-dotenv
apscheduler
requests
